backend/agentiq/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcivic-agentiq"
version = "0.1.0"
description = "AgentCivic AgentIQ workflow integration"
dependencies = ["nvidia-nat[langchain]"]

[project.entry-points.'nat.plugins']
agentcivic_agentiq = "agentcivic_agentiq.register"
